backend/requirements.txt
fastapi
uvicorn
python-multipart
pydicom
numpy
Pillow
torch
torchvision
pydantic
